Moved getWindDirection -functionality to helpers
