+++ src/components/weather/NextHourForecastPanel.tsx
@@ -17,7 +17,7 @@ import {
 import { selectTimeZone } from '@store/location/selector';
 import { weatherSymbolGetter } from '@assets/images';
 
-import { getFeelsLikeIconName } from '@utils/helpers';
+import { getFeelsLikeIconName, getWindDirection } from '@utils/helpers';
 import { CustomTheme, GRAY_1 } from '@utils/colors';
 
 import Icon from '@components/common/Icon';
@@ -115,35 +115,6 @@ const NextHourForecastPanel: React.FC<NextHourForecastPanelProps> = ({
     precipitationUnit,
     nextHourForecast.precipitation1h
   );
-
-  // wind direction can be with either one degree accuracy or uses cardinals (N,NE,E,...)
-  const useCardinals = Config.get('weather').useCardinalsForWindDirection;
-  let direction = 0;
-  if (useCardinals) {
-    const tempDir = nextHourForecast.windDirection || 0;
-    if ((tempDir >= 338 && tempDir <= 360) || (tempDir >= 0 && tempDir <= 22)) {
-      direction = 0; // N
-    } else if (tempDir >= 23 && tempDir <= 67) {
-      direction = 45; // NE
-    } else if (tempDir >= 68 && tempDir <= 112) {
-      direction = 90; // E
-    } else if (tempDir >= 113 && tempDir <= 157) {
-      direction = 135; // SE
-    } else if (tempDir >= 158 && tempDir <= 202) {
-      direction = 180; // S
-    } else if (tempDir >= 203 && tempDir <= 247) {
-      direction = 225; // SW
-    } else if (tempDir >= 248 && tempDir <= 292) {
-      direction = 270; // W
-    } else if (tempDir >= 293 && tempDir <= 337) {
-      direction = 315; // NW
-    }
-    // for some reason icon is pointing NW instead of N => +45
-    // wind value needs 180 degree switch to show correctly where wind is coming from
-    direction = direction + 45 - 180;
-  } else {
-    direction = (nextHourForecast.windDirection || 0) + 45 - 180;
-  }
 
   return (
     <View style={styles.container}>
@@ -249,7 +220,9 @@ const NextHourForecastPanel: React.FC<NextHourForecastPanelProps> = ({
                 {
                   transform: [
                     {
-                      rotate: `${direction}deg`,
+                      rotate: `${getWindDirection(
+                        nextHourForecast.windDirection
+                      )}deg`,
                     },
                   ],
                 },
